remove duplicated cache key and reorganize rsc entry
