+++ packages/next/client/rsc/index.ts
@@ -1,0 +1,1 @@
+export { useRefreshRoot as unstable_useRefreshRoot } from './refresh'
