+++ packages/next/rsc.d.ts
@@ -1,4 +1,1 @@
-export {
-  useRefreshRoot as unstable_useRefreshRoot,
-  PageNotFound,
-} from './dist/client/rsc'
+export * from './dist/client/rsc'
